test(segment): migrate modes
